test: migrate part of Mentions tests
